pyproject.toml
[project]
name = "arbos"
version = "0.1.0"
description = "Discord-driven coding agent"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "python-dotenv",
    "discord.py>=2.3",
    "httpx",
    "fastapi",
    "uvicorn",
    "cryptography",
]

[project.scripts]
arbos = "arbos.main:main"

[tool.setuptools]
packages = ["arbos"]
